requirements.txt
flask
openai
python-docx
pdfplumber
nltk
